Completed management of methods that return arrays.
The Collection Facet and its collaborator TypeOf Facet characterize a
cloned specification of 'array' that is stored as the return type of the
method. A new NakedObject is created via Prototype pattern when the
method is called, and new NakedObjects are created in the same manner
from the Facets to navigate the Collection.
The creation of the Facets is addressed by the ProgModelFactory, which
should also be refactored to include the Invocation Facet creation.

NakedBareObject::__toString() now recognizes the presence of Collection\
Facets to produce a representation like '5 stdClass'. The DisplayObject
view helper does the same but producing a table representation.
Minor refactorings done on method names in NakedPhp\Reflect.
Added stdClass to recognized native types.