Added extraction of object from an array to view and work with them as standalone ones.
The Collection Facet interface and its CollectionArray implementation
have been expanded with the toArray(NakedObject) method.
The controller accepts the request to view a field of a collection
and use the Facet to extract it and put it in the session independently.
The DisplayObject view helper adds a link to trigger this action.