Added selection of object parameters for method calls via html select.
Refactoring of StateManager: scalar values are wrapped while recognized
objects (via a key) are extracted from the session. Elimination of
entity-related methods.
Refactoring of Invocation Facet: parameters should be NakedObject
instances.
Refactoring of NakedFactory: wraps also scalar values that may be passed
in a form, such as strings and integers.